+++ educa100/src/main/java/com/educa100/controller/dto/response/CursoResponse.java
@@ -1,4 +1,4 @@
 package com.educa100.controller.dto.response;
 
-public record CursoResponse(int id_curso) {
+public record CursoResponse(Long id_curso) {
 }
