+++ educa100/src/main/java/com/educa100/controller/dto/response/DocenteResponse.java
@@ -1,4 +1,4 @@
 package com.educa100.controller.dto.response;
 
-public record DocenteResponse(int id_docente) {
+public record DocenteResponse(Long id_docente) {
 }
