+++ educa100/src/main/java/com/educa100/controller/dto/response/TurmaResponse.java
@@ -1,4 +1,4 @@
 package com.educa100.controller.dto.response;
 
-public record TurmaResponse(int id_turma) {
+public record TurmaResponse(Long id_turma) {
 }
