+++ educa100/src/main/java/com/educa100/service/NotaService.java
@@ -7,9 +7,9 @@ import java.util.List;
 
 public interface NotaService {
     public NotaEntity salvar(NotaEntity nota);
-    public void atualizar(int id);
-    public void removerPorId(int id);
-    public List<NotaEntity> listarPorIdAluno(int id);
+    public void atualizar(Long id);
+    public void removerPorId(Long id);
+    public List<NotaEntity> listarPorIdAluno(Long id);
     public List<NotaEntity> listarTodos();
-    public NotaEntity listarPorId(int id);
+    public NotaEntity listarPorId(Long id);
 }
